+++ internal/aghtest/os_others.go
@@ -1,0 +1,13 @@
+// +build !windows
+
+package aghtest
+
+import (
+	"testing"
+)
+
+func prepareTestDir(t *testing.T) (dir string) {
+	t.Helper()
+
+	return t.TempDir()
+}