+++ internal/tools/doc.go
@@ -1,0 +1,5 @@
+// Package tools and its main module are a nested internal module containing our
+// development tool dependencies.
+//
+// See https://github.com/golang/go/wiki/Modules#how-can-i-track-tool-dependencies-for-a-module.
+package tools