+++ scripts/make/build-release.sh
@@ -121,7 +121,6 @@ readonly mipses='softfloat'
 
 #    os  arch      arm mips       snap
 readonly platforms="\
-darwin   386       0   0          0
 darwin   amd64     0   0          0
 freebsd  386       0   0          0
 freebsd  amd64     0   0          0